轨迹点上传

+++ internal-common/src/main/java/com/dong/internalcommon/request/PointsDTO.java
@@ -2,9 +2,10 @@ package com.dong.internalcommon.request;
 
 import lombok.Data;
 
+
 @Data
 public class PointsDTO {
 
     private String location;
-    private Integer locatetime;
+    private Long locatetime;
 }
